Show errors during static page generation
Previously when errors were thrown during static page rendering, they
were swallowed by static-site-generator-webpack-plugin and the page was
rendered with blank content.

This commit adds a check to stats.hasErrors() and returns the error
string from stats to the callback. However, the stats object does not
expose the original errors, so we also print the stack trace to make
it possible to debug errors only thrown during server rendering.