[www] feat(starters-showcase): Improve performance (#9037)
Hi, Gatsby community! 👋 

I love the new design of the showcase for Gatsby starter but one thing I noticed was the bad performance when trying to filter using the search input on the right. The problem was that every time the user type on the input, all the Gatsby starter is filtered.

The solution I made is to add some `debounce` magic on every `onChange` and also move the value of the input to the `state`. Now on every keystroke, the application re-render only the `input` and after 500 ms has passed after the `onChange` call I filter the starter list.

That's all folks, hope it helps! 😄