refactor(gatsby): remove parsePath from gatsby core, and move to gatsby-link (#9957)
* refactor: remove parsePath from gatsby core, and move to gatsby-link

Moving a utility function (that was not used!) in gatsby core into where
it belongs, gatsby-link. This was discovered via a community member
(@travi), as he was attempting to use storybook to test a Link component
(outside of Gatsby), e.g. `import Link from 'gatsby-link'`, which was
then requiring `gatsby` core.

This also generally solves some weirdness that the cyclical require was
kinda weird, e.g. gatsby -> gatsby-link -> gatsby

I do not believe this is a breaking change, since the version required
of gatsby-link will just be bumped by gatsby, and the export is still
exported from the top level gatsby module

* test(gatsby-link): add additional unit tests for parse path

* chore: use test blocks

* chore: fix crlf

* chore: import parsePath from gatsby-link

* chore: export parsePath (whooops)

* chore: remove compiled file