remove tailwind.js as dependency in postcss config (#14087)
When following this guide, I including the pointer to `tailwind.js` in the postcss config, but it broke when I ran `gatsby develop`. I checked the tailwind site, and they do not include `tailwind.js` in the postcss config [here](https://tailwindcss.com/docs/installation/#using-tailwind-with-postcss). Not sure if this was a change caused by the recent tailwind release, but removing it and running `gatsby develop` seems to work.