docs: Slight update to surge deployment instructions (#17855)
Docs are currently specifying to use `surge` after installing the Surge
CLI to create an account, but after account creation it continues with
steps for deploying to surge. It is too soon to deploy at this point in
the tutorial because `gatsby build` has not been run yet. Using `surge
login` instead, it will end after the user creates an account by logging
in with email and password.