feat(gatsby): enable Jobs API v2 (#19858)
Added a new action called createJobV2 to support the new api. createJob is still available to keep backward-compatibility. I'll add a deprecation message when creatJobV2 is fully operational.

The createJobV2 needs a job object that contains 4 arguments name, inputPaths, outputDir & args. These args are used to create a unique content digest to make sure a job is deterministic.

InputPaths are converted into relative paths before sending it to the worker as they need to be filesystem agnostic.

More info on why can be found in #19831